+++ src/RcsbCollectTools/PropertyCollector/GroupPropertyCollector.ts
@@ -1,5 +1,5 @@
 import {RcsbClient} from "../../RcsbGraphQL/RcsbClient";
-import {CoreGroup, QueryGroupArgs} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Yosemite/GqlTypes";
+import {GroupPolymerEntity, QueryPolymer_Entity_GroupArgs} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Yosemite/GqlTypes";
 
 export interface GroupPropertyInterface{
     groupName: string;
@@ -10,9 +10,9 @@ export class GroupPropertyCollector {
 
     private readonly rcsbFvQuery: RcsbClient = new RcsbClient();
 
-    public async collect(requestConfig: QueryGroupArgs): Promise<GroupPropertyInterface> {
+    public async collect(requestConfig: QueryPolymer_Entity_GroupArgs): Promise<GroupPropertyInterface> {
         try {
-            const result: CoreGroup = await this.rcsbFvQuery.requestGroupInfo(requestConfig);
+            const result: GroupPolymerEntity = await this.rcsbFvQuery.requestGroupInfo(requestConfig);
             return {
                 groupName: result.rcsb_group_info.group_name,
                 groupDescription: result.rcsb_group_info.group_description