+++ src/RcsbCollectTools/Translators/GroupMemberCollector.ts
@@ -1,5 +1,5 @@
 import {RcsbClient} from "../../RcsbGraphQL/RcsbClient";
-import {CoreGroup, QueryGroupArgs} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Yosemite/GqlTypes";
+import {GroupPolymerEntity, QueryPolymer_Entity_GroupArgs} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Yosemite/GqlTypes";
 import {PolymerEntityInstanceInterface} from "./PolymerEntityInstancesCollector";
 import {MultipleEntityInstancesCollector} from "./MultipleEntityInstancesCollector";
 
@@ -7,9 +7,9 @@ export class GroupMemberCollector {
 
     private readonly rcsbFvQuery: RcsbClient = new RcsbClient();
 
-    public async collect(requestConfig: QueryGroupArgs): Promise<Array<PolymerEntityInstanceInterface>> {
+    public async collect(requestConfig: QueryPolymer_Entity_GroupArgs): Promise<Array<PolymerEntityInstanceInterface>> {
         try {
-            const result: CoreGroup = await this.rcsbFvQuery.requestGroupInfo(requestConfig);
+            const result: GroupPolymerEntity = await this.rcsbFvQuery.requestGroupInfo(requestConfig);
             if(result.rcsb_group_info.group_members_granularity === "polymer_entity"){
                 const multipleEntityInstancesCollector: MultipleEntityInstancesCollector = new MultipleEntityInstancesCollector();
                 return  await multipleEntityInstancesCollector.collect({entity_ids: result.rcsb_group_container_identifiers.group_member_ids});