+++ src/RcsbGraphQL/RcsbQueryGroup.ts
@@ -1,20 +1,19 @@
 import query from "./Queries/Yosemite/QueryGroup.graphql";
-import {CoreGroup, QueryGroupArgs} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Yosemite/GqlTypes";
+import {GroupPolymerEntity, QueryPolymer_Entity_GroupArgs} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/Types/Yosemite/GqlTypes";
 import {RcsbCoreQueryInterface} from "./RcsbCoreQueryInterface";
 import {GraphQLRequest} from "@rcsb/rcsb-api-tools/build/RcsbGraphQL/GraphQLRequest";
 
 interface GroupResultInterface {
-    group: CoreGroup;
+    group: GroupPolymerEntity;
 }
 
-export class RcsbQueryGroup implements RcsbCoreQueryInterface<QueryGroupArgs,CoreGroup>{
+export class RcsbQueryGroup implements RcsbCoreQueryInterface<QueryPolymer_Entity_GroupArgs,GroupPolymerEntity>{
     readonly client: GraphQLRequest = new GraphQLRequest("yosemite");
-    public async request(requestConfig: QueryGroupArgs): Promise<CoreGroup> {
+    public async request(requestConfig: QueryPolymer_Entity_GroupArgs): Promise<GroupPolymerEntity> {
         try {
-            const response:GroupResultInterface = await this.client.request<QueryGroupArgs, GroupResultInterface>(
+            const response:GroupResultInterface = await this.client.request<QueryPolymer_Entity_GroupArgs, GroupResultInterface>(
                 {
-                    group_id: requestConfig.group_id,
-                    aggregation_method: requestConfig.aggregation_method
+                    group_id: requestConfig.group_id
                 },
                 query
             );